Add media management security evaluation and list media info methods.

+++ components/org.wso2.carbon.identity.media.core/src/main/java/org/wso2/carbon/identity/media/core/file/FileBasedStorageSystemImpl.java
@@ -31,6 +31,8 @@ import org.wso2.carbon.identity.media.core.StorageSystem;
 import org.wso2.carbon.identity.media.core.exception.StorageSystemException;
 import org.wso2.carbon.identity.media.core.model.FileSecurity;
 import org.wso2.carbon.identity.media.core.model.MediaFileDownloadData;
+import org.wso2.carbon.identity.media.core.model.MediaInformation;
+import org.wso2.carbon.identity.media.core.model.MediaInformationMetadata;
 import org.wso2.carbon.identity.media.core.model.MediaMetadata;
 
 import java.io.File;
@@ -56,18 +58,20 @@ import java.util.List;
 
 import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.CONFIGURABLE_UPLOAD_LOCATION;
 import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.FILE_CREATION_TIME_ATTRIBUTE;
-import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.FILE_SECURITY_ALLOWED_ALL;
-import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.FILE_SECURITY_ALLOWED_USERS;
+import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.MEDIA_CONTENT_TYPE;
+import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.MEDIA_CREATED_TIME;
+import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.MEDIA_FILE_LAST_ACCESSED_TIME;
+import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.MEDIA_NAME;
+import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.MEDIA_RESOURCE_OWNER;
+import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.MEDIA_SECURITY;
+import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.MEDIA_SECURITY_ALLOWED_ALL;
+import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.MEDIA_SECURITY_ALLOWED_USERS;
 import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.MEDIA_STORE;
-import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.METADATA_FILE_CONTENT_TYPE;
-import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.METADATA_FILE_CREATED_TIME;
+import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.MEDIA_TAG;
 import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.METADATA_FILE_EXTENSION;
-import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.METADATA_FILE_LAST_ACCESSED_TIME;
-import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.METADATA_FILE_NAME;
-import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.METADATA_FILE_RESOURCE_OWNER;
-import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.METADATA_FILE_SECURITY;
 import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.METADATA_FILE_SUFFIX;
-import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.METADATA_FILE_TAG;
+import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.PROTECTED_DOWNLOAD_ACCESS;
+import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.PUBLIC_DOWNLOAD_ACCESS;
 import static org.wso2.carbon.identity.media.core.util.StorageSystemConstants.SYSTEM_PROPERTY_CARBON_HOME;
 
 /**
@@ -158,15 +162,58 @@ public class FileBasedStorageSystemImpl implements StorageSystem {
     }
 
     @Override
-    public void deleteFile(String id, String type, String tenantDomain) throws StorageSystemException {
+    public boolean evaluateMediaManagementSecurityForEndUser(String id, String type, String tenantDomain) throws
+            StorageSystemException {
+
+        File file;
+        try {
+            file = getMediaMetadataFile(id, type, tenantDomain);
+            if (file != null && file.exists()) {
+                return isMediaManagementAllowed(file);
+            }
+            return false;
+        } catch (IOException e) {
+            String errorMsg = String.format("Error while retrieving metadata for stored file with id: %s and of type " +
+                    "%s in tenant domain: %s", id, type, tenantDomain);
+            throw new StorageSystemException(errorMsg, e);
+        } catch (ParseException e) {
+            String errorMsg = String.format("Unable to parse metadata in JSON format for stored file with id : %s " +
+                    "and of type %s in tenant domain: %s", id, type, tenantDomain);
+            throw new StorageSystemException(errorMsg, e);
+        }
+    }
+
+    @Override
+    public boolean isMediaDeleted(String id, String type, String tenantDomain) throws StorageSystemException {
 
         try {
-            deleteMediaFile(id, type, tenantDomain);
+            return isFileDeleted(id, type, tenantDomain);
         } catch (IOException e) {
             String errorMsg = String.format("Error while deleting the stored file of type %s.", type);
             throw new StorageSystemException(errorMsg, e);
         }
+    }
 
+    @Override
+    public MediaInformation getMediaInformation(String id, String type, String tenantDomain) throws
+            StorageSystemException {
+
+        File file;
+        try {
+            file = getMediaMetadataFile(id, type, tenantDomain);
+            if (file != null && file.exists()) {
+                return getMediaInformation(file, type, id);
+            }
+            return null;
+        } catch (IOException e) {
+            String errorMsg = String.format("Error while retrieving metadata for stored file with id: %s and of type " +
+                    "%s in tenant domain: %s", id, type, tenantDomain);
+            throw new StorageSystemException(errorMsg, e);
+        } catch (ParseException e) {
+            String errorMsg = String.format("Unable to parse metadata in JSON format for stored file with id : %s " +
+                    "and of type %s in tenant domain: %s", id, type, tenantDomain);
+            throw new StorageSystemException(errorMsg, e);
+        }
     }
 
     @Override
@@ -276,7 +323,7 @@ public class FileBasedStorageSystemImpl implements StorageSystem {
                 StandardCharsets.UTF_8)) {
             JSONParser parser = new JSONParser();
             JSONObject metadata = (JSONObject) parser.parse(reader);
-            metadata.put(METADATA_FILE_LAST_ACCESSED_TIME, timeStamp);
+            metadata.put(MEDIA_FILE_LAST_ACCESSED_TIME, timeStamp);
 
             try (FileOutputStream fileStream = new FileOutputStream(metadataFile);
                  Writer writer = new OutputStreamWriter(fileStream, StandardCharsets.UTF_8)) {
@@ -291,7 +338,7 @@ public class FileBasedStorageSystemImpl implements StorageSystem {
                 StandardCharsets.UTF_8)) {
             JSONParser parser = new JSONParser();
             JSONObject metadata = (JSONObject) parser.parse(reader);
-            return (String) metadata.get(METADATA_FILE_CONTENT_TYPE);
+            return (String) metadata.get(MEDIA_CONTENT_TYPE);
         }
     }
 
@@ -312,9 +359,9 @@ public class FileBasedStorageSystemImpl implements StorageSystem {
         try (InputStreamReader reader = new InputStreamReader(new FileInputStream(file), StandardCharsets.UTF_8)) {
             JSONParser jsonParser = new JSONParser();
             Object metadata = jsonParser.parse(reader);
-            HashMap fileSecurity = (HashMap) ((JSONObject) metadata).get(METADATA_FILE_SECURITY);
+            HashMap fileSecurity = (HashMap) ((JSONObject) metadata).get(MEDIA_SECURITY);
             if (fileSecurity != null) {
-                return (Boolean) fileSecurity.get(FILE_SECURITY_ALLOWED_ALL);
+                return (Boolean) fileSecurity.get(MEDIA_SECURITY_ALLOWED_ALL);
             }
         }
         return false;
@@ -325,9 +372,9 @@ public class FileBasedStorageSystemImpl implements StorageSystem {
         try (InputStreamReader reader = new InputStreamReader(new FileInputStream(file), StandardCharsets.UTF_8)) {
             JSONParser jsonParser = new JSONParser();
             Object metadata = jsonParser.parse(reader);
-            HashMap fileSecurity = (HashMap) ((JSONObject) metadata).get(METADATA_FILE_SECURITY);
+            HashMap fileSecurity = (HashMap) ((JSONObject) metadata).get(MEDIA_SECURITY);
             if (fileSecurity != null) {
-                if ((Boolean) fileSecurity.get(FILE_SECURITY_ALLOWED_ALL)) {
+                if ((Boolean) fileSecurity.get(MEDIA_SECURITY_ALLOWED_ALL)) {
                     return true;
                 }
                 return isUserAllowed(fileSecurity);
@@ -336,9 +383,20 @@ public class FileBasedStorageSystemImpl implements StorageSystem {
         }
     }
 
+    private boolean isMediaManagementAllowed(File file) throws IOException, ParseException {
+
+        try (InputStreamReader reader = new InputStreamReader(new FileInputStream(file), StandardCharsets.UTF_8)) {
+            JSONParser jsonParser = new JSONParser();
+            Object metadata = jsonParser.parse(reader);
+            String resourceOwner = (String) ((JSONObject) metadata).get(MEDIA_RESOURCE_OWNER);
+            return StringUtils.isNotBlank(resourceOwner) &&
+                    resourceOwner.equals(PrivilegedCarbonContext.getThreadLocalCarbonContext().getUsername());
+        }
+    }
+
     private boolean isUserAllowed(HashMap fileSecurity) {
 
-        ArrayList allowedUsers = (ArrayList) fileSecurity.get(FILE_SECURITY_ALLOWED_USERS);
+        ArrayList allowedUsers = (ArrayList) fileSecurity.get(MEDIA_SECURITY_ALLOWED_USERS);
         if (allowedUsers != null) {
             for (Object allowedUser : allowedUsers) {
                 if (allowedUser instanceof String) {
@@ -352,6 +410,43 @@ public class FileBasedStorageSystemImpl implements StorageSystem {
         return false;
     }
 
+    private MediaInformation getMediaInformation(File file, String type, String id) throws IOException, ParseException {
+
+        MediaInformation mediaInformation = new MediaInformation();
+        MediaInformationMetadata mediaInformationMetadata = new MediaInformationMetadata();
+        try (InputStreamReader reader = new InputStreamReader(new FileInputStream(file), StandardCharsets.UTF_8)) {
+            JSONParser jsonParser = new JSONParser();
+            Object metadata = jsonParser.parse(reader);
+
+            String tag = (String) ((JSONObject) metadata).get(MEDIA_TAG);
+            mediaInformationMetadata.setTag(tag);
+
+            HashMap fileSecurityMap = (HashMap) ((JSONObject) metadata).get(MEDIA_SECURITY);
+            boolean allowedAll = (Boolean) fileSecurityMap.get(MEDIA_SECURITY_ALLOWED_ALL);
+            List<String> allowedUsers = (List<String>) fileSecurityMap.get(MEDIA_SECURITY_ALLOWED_USERS);
+            FileSecurity fileSecurity;
+            if (CollectionUtils.isNotEmpty(allowedUsers)) {
+                fileSecurity = new FileSecurity(allowedAll, allowedUsers);
+            } else {
+                fileSecurity = new FileSecurity(allowedAll);
+            }
+            mediaInformationMetadata.setSecurity(fileSecurity);
+
+            ArrayList<String> links = new ArrayList<>();
+            String access;
+            if (allowedAll) {
+                access = PUBLIC_DOWNLOAD_ACCESS;
+            } else {
+                access = PROTECTED_DOWNLOAD_ACCESS;
+            }
+            links.add(String.format("/%s/%s/%s", access, type, id));
+            mediaInformation.setLinks(links);
+
+            mediaInformation.setMediaInformationResponseMetadata(mediaInformationMetadata);
+        }
+        return mediaInformation;
+    }
+
     private void storeMediaMetadata(Path targetLocation, String fileName, String fileContentType, String fileTag,
                                     String resourceOwner, FileSecurity fileSecurity, String timestamp)
             throws IOException {
@@ -362,20 +457,20 @@ public class FileBasedStorageSystemImpl implements StorageSystem {
         JSONObject metadata = new JSONObject();
 
         if (StringUtils.isNotBlank(fileName)) {
-            metadata.put(METADATA_FILE_NAME, fileName);
+            metadata.put(MEDIA_NAME, fileName);
         }
         if (StringUtils.isNotBlank(fileContentType)) {
-            metadata.put(METADATA_FILE_CONTENT_TYPE, fileContentType);
+            metadata.put(MEDIA_CONTENT_TYPE, fileContentType);
         }
         if (StringUtils.isNotBlank(fileTag)) {
-            metadata.put(METADATA_FILE_TAG, fileTag);
+            metadata.put(MEDIA_TAG, fileTag);
         }
         if (StringUtils.isNotBlank(resourceOwner)) {
-            metadata.put(METADATA_FILE_RESOURCE_OWNER, resourceOwner);
+            metadata.put(MEDIA_RESOURCE_OWNER, resourceOwner);
         }
         storeFileSecurityMetadata(fileSecurity, metadata);
-        metadata.put(METADATA_FILE_CREATED_TIME, timestamp);
-        metadata.put(METADATA_FILE_LAST_ACCESSED_TIME, timestamp);
+        metadata.put(MEDIA_CREATED_TIME, timestamp);
+        metadata.put(MEDIA_FILE_LAST_ACCESSED_TIME, timestamp);
 
         try (FileOutputStream fileStream = new FileOutputStream(metadataTargetLocation.toFile());
              Writer writer = new OutputStreamWriter(fileStream, StandardCharsets.UTF_8)) {
@@ -388,18 +483,18 @@ public class FileBasedStorageSystemImpl implements StorageSystem {
         JSONObject fileSecurityJSON = new JSONObject();
 
         boolean allowedAll = fileSecurity.isAllowedAll();
-        fileSecurityJSON.put(FILE_SECURITY_ALLOWED_ALL, allowedAll);
+        fileSecurityJSON.put(MEDIA_SECURITY_ALLOWED_ALL, allowedAll);
 
         if (!allowedAll) {
             List<String> allowedUsers = fileSecurity.getAllowedUsers();
             if (CollectionUtils.isNotEmpty(allowedUsers)) {
-                fileSecurityJSON.put(FILE_SECURITY_ALLOWED_USERS, allowedUsers);
+                fileSecurityJSON.put(MEDIA_SECURITY_ALLOWED_USERS, allowedUsers);
             }
         }
-        metadata.put(METADATA_FILE_SECURITY, fileSecurityJSON);
+        metadata.put(MEDIA_SECURITY, fileSecurityJSON);
     }
 
-    private void deleteMediaFile(String id, String type, String tenantDomain) throws IOException {
+    private boolean isFileDeleted(String id, String type, String tenantDomain) throws IOException {
 
         int tenantId = IdentityTenantUtil.getTenantId(tenantDomain);
 
@@ -408,9 +503,13 @@ public class FileBasedStorageSystemImpl implements StorageSystem {
             Path filePath = fileStorageLocation.resolve(id).normalize();
             String metadataFileName = id + METADATA_FILE_SUFFIX + METADATA_FILE_EXTENSION;
             Path metadataFilePath = fileStorageLocation.resolve(metadataFileName).normalize();
-            Files.deleteIfExists(filePath);
-            Files.deleteIfExists(metadataFilePath);
+            if (Files.exists((filePath))) {
+                Files.delete(filePath);
+                Files.deleteIfExists(metadataFilePath);
+                return true;
+            }
         }
+        return false;
     }
 
     private Path getStorageDirectory(String fileType, int tenantId, String id) {
